style: 디렉토리 병합
