--- src/main/java/com/petstagram/auth/dto/LoginRequestDto.java
@@ -1,20 +1,0 @@
-package com.petstagram.auth.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import lombok.Getter;
-
-@Getter
-public class LoginRequestDto {
-
-    @Pattern(regexp = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,6}$", message = "올바른 이메일 형식을 입력해 주세요")
-    private final String email;
-
-    @NotBlank // 패턴 논의 전
-    private final String password;
-
-    public LoginRequestDto(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-}
