--- src/main/java/com/petstagram/auth/dto/SignUpRequestDto.java
@@ -1,26 +1,0 @@
-package com.petstagram.auth.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import lombok.Getter;
-
-@Getter
-public class SignUpRequestDto {
-
-    @NotBlank(message = "올바른 이름을 입력해주세요")
-    private final String name;
-
-    @Pattern(regexp = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,6}$", message = "올바른 이메일 형식을 입력해 주세요")
-    private final String email;
-
-    // 패턴 논의되지 않았음, 우선 보류
-    @NotBlank
-    private final String password;
-
-    public SignUpRequestDto(String name, String email, String password) {
-        this.name = name;
-        this.email = email;
-        this.password = password;
-    }
-
-}
