+++ src/main/java/com/petstagram/model/entity/User.java
@@ -22,7 +22,7 @@ public class User extends Time{
     @Column(nullable = false)
     private String name;
 
-    @Column(nullable = false)
+    @Column(nullable = false, unique = true)
     private String email;
 
     @Column(nullable = false)
